Refactore NavbarComp and ProjectsComp components
Refactored the NavbarComp component to add a new import statement for the `usePathname` hook from "next/navigation". Also added a new variable `pathname` to store the current pathname using the `usePathname` hook.

Refactored the ProjectsComp component to replace the div element with a Card component from "@nextui-org/react". Updated the className to include a background color and adjusted the styling accordingly.